pyproject.toml
[project]
name = "yavdr-frontend"
version = "0.1.0"
description = "controls the frontend on a yavdr-ansible installation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.12.0",
    "pygobject>=3.52.3",
    "ruamel-yaml>=0.18.12",
    "sdbus>=0.14.0",
]

[project.scripts]
frontend-dbus-send = "yavdr_frontend:frontend_dbus_send.main"
start-desktop = "yavdr_frontend:start_desktop.main"
run-desktop = "yavdr_frontend:run_desktop.main"
yavdr-frontend = "yavdr_frontend:startup.main"
yavdr-frontendctl = "yavdr_frontend:frontendctl.main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["pygobject-stubs>=2.13.0"]
